% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_plot.R
\name{ci_plot}
\alias{ci_plot}
\title{Custom Cumulative incidence plot with risk table}
\usage{
ci_plot(
  dat,
  time,
  status,
  variable,
  title,
  xlab,
  note,
  pallette,
  legend_labs,
  xlim,
  ylim,
  break.time.by,
  breaks = NULL
)
}
\arguments{
\item{dat}{data frame}

\item{time}{variable, in quotes, with the survival time information}

\item{status}{variable, in quotes, with crr survival status}

\item{variable}{variable, in quotes, of interest to compare}

\item{title}{title of graph}

\item{xlab}{x-axis label}

\item{legend_labs}{vector}

\item{xlim}{vector}

\item{ylim}{vector}

\item{break.time.by}{vector}

\item{palette}{color palette}
}
\description{
Hodgepodge of surviminer, survival, and custom code to get a
cumulative incidence curve with risk table the way I like it
}
\examples{
# example add
}
\author{
Margaret L. Hannum
}
