% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funnum-package.R
\docType{package}
\name{funnum-package}
\alias{funnum}
\alias{funnum-package}
\title{funnum: Helper functions to get projects done}
\description{
funnum is a collection of functions from other packages, customized for my purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/margarethannum/funnum}
  \item Report bugs at \url{https://github.com/margarethannum/funnum/issues}
}

}
\author{
\strong{Maintainer}: Margaret Hannum \email{hannumm@mskcc.org}

}
\keyword{internal}
