% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcomprisk_mh.R
\name{ggcomprisk_mh}
\alias{ggcomprisk_mh}
\title{Competing risks modified function (from survminer)}
\usage{
ggcomprisk_mh(fit, gnames = NULL, gsep = " ", multiple_panels = TRUE)
}
\arguments{
\item{fit}{cuminc model}
}
\description{
Make color be group instead of event
}
\author{
Margaret L. Hannum
}
