% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cuminc_w_risktab.R
\name{plot_cuminc_w_risktab}
\alias{plot_cuminc_w_risktab}
\title{Plot Cumulative Incidence with Survminer Risk Table}
\usage{
plot_cuminc_w_risktab(
  title,
  note,
  variable,
  survtime,
  survstatus_gray,
  survstatus_reg,
  dat,
  legend_labs = levels(as.factor(dat[[variable]])),
  xlim,
  ylim,
  break.time.by,
  breaks
)
}
\arguments{
\item{title}{character title}

\item{note}{if you want to annotate}

\item{variable}{group variable}

\item{survtime}{survival time variable}

\item{survstatus_gray}{survival status variable for competing risks (0, 1, 2)}

\item{survstatus_reg}{survival status variable for standard kaplan meier (0,1): required for risk table}

\item{dat}{dataframe}

\item{legend_labs}{Vector of group factors. Default is levels(as.factor(dat[[variable]]))}

\item{xlim}{vector of x limits}

\item{ylim}{vector of y limits}

\item{break.time.by}{numeric}

\item{breaks}{numeric vector}
}
\description{
Pastes together a modified version of ggcompetingrisks with a
custom ggrisktable
}
\examples{
ADD EXAMPLE
}
\seealso{
\code{\link[survminer]{ggcompetingrisks}}, \code{\link[survminer]{ggrisktable}}, \code{\link[cmprsk]{cuminc}}
}
\author{
Margaret L. Hannum
}
