% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcomprisk_mh.R
\name{ggcomprisk_mh}
\alias{ggcomprisk_mh}
\title{Competing risks modified function (from survminer)}
\usage{
ggcomprisk_mh(
  fit,
  gnames = NULL,
  gsep = " ",
  multiple_panels = TRUE,
  stratified = TRUE
)
}
\arguments{
\item{fit}{cuminc model}

\item{multiple_panels}{logical input (default = TRUE), whether to split stratification into two panels}

\item{stratified}{logical input (default = TRUE), whether input model is stratified or not}
}
\description{
Make color be group instead of event
}
\author{
Margaret L. Hannum
}
