% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggglmnet.R
\name{ggglmnet}
\alias{ggglmnet}
\title{Plot nice lasso results from glmnet}
\usage{
ggglmnet(fit, beta, intercept)
}
\arguments{
\item{fit}{object of class `glmnet`}

\item{beta}{class dgCMatrix from extracting coefficients from model fit}

\item{intercept}{string indicating name of intercept in model results}
}
\description{
Labels
}
\examples{


}
\seealso{
\code{\link[survminer]{ggcompetingrisks}}
}
\author{
Margaret L. Hannum
}
