% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtab_rowcol.R
\name{xtab_rowcol}
\alias{xtab_rowcol}
\title{Calculate row AND column crosstabulatd percent}
\usage{
xtab_rowcol(x, rowvar, colvar)
}
\arguments{
\item{x}{crosstabulated matrix (e.g. cluster assignments)}

\item{rowvar}{name of row variables}

\item{colvar}{name of column variables}
}
\description{
Useful in comparing cluster assignments
}
\author{
Margaret L. Hannum
}
